New formalization.
