+++ sdk/nodejs/s3/s3Mixins.ts
@@ -12,11 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Re-export the awssdk.S3 class/namespace out as 'runtime' so there is easy access to this
-// API at actual cloud-runtime.
-type awsSdkType = typeof import("aws-sdk");
-export const runtime: awsSdkType["S3"] = <any>undefined;
-
 import * as pulumi from "@pulumi/pulumi"
 import { Bucket } from "./bucket";
 import { BucketNotification } from "./bucketNotification";
